@@ -13,49 +13,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~10.0.6",
-    "@angular/core": "~10.0.6",
-    "@angular/forms": "~10.0.6",
-    "@angular/platform-browser": "~10.0.6",
-    "@angular/platform-browser-dynamic": "~10.0.6",
-    "@angular/router": "~10.0.6",
-    "@ionic-native/core": "^5.27.0",
-    "@ionic-native/splash-screen": "^5.27.0",
-    "@ionic-native/status-bar": "^5.27.0",
-    "@ionic/angular": "^5.3.1",
-    "chart.js": "^2.9.3",
+    "@angular/common": "~11.0.2",
+    "@angular/core": "~11.0.2",
+    "@angular/forms": "~11.0.2",
+    "@angular/platform-browser": "~11.0.2",
+    "@angular/platform-browser-dynamic": "~11.0.2",
+    "@angular/router": "~11.0.2",
+    "@ionic-native/core": "^5.30.0",
+    "@ionic-native/splash-screen": "^5.30.0",
+    "@ionic-native/status-bar": "^5.30.0",
+    "@ionic/angular": "^5.5.1",
+    "chart.js": "^2.9.4",
     "chartjs-plugin-zoom": "^0.7.7",
-    "core-js": "^3.6.5",
-    "ng2-charts": "^2.3.2",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "core-js": "^3.8.0",
+    "ng2-charts": "^2.4.2",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1000.4",
-    "@angular-devkit/build-angular": "~0.1000.4",
-    "@angular-devkit/core": "~10.0.4",
-    "@angular-devkit/schematics": "~10.0.4",
-    "@angular/cli": "~10.0.4",
-    "@angular/compiler": "~10.0.6",
-    "@angular/compiler-cli": "~10.0.6",
-    "@angular/language-service": "~10.0.6",
-    "@ionic/angular-toolkit": "^2.3.0",
-    "@types/jasmine": "~3.5.11",
+    "@angular-devkit/architect": "~0.1100.2",
+    "@angular-devkit/build-angular": "~0.1100.2",
+    "@angular-devkit/core": "~11.0.2",
+    "@angular-devkit/schematics": "~11.0.2",
+    "@angular/cli": "~11.0.2",
+    "@angular/compiler": "~11.0.2",
+    "@angular/compiler-cli": "~11.0.2",
+    "@angular/language-service": "~11.0.2",
+    "@ionic/angular-toolkit": "^3.0.0",
+    "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~14.0.27",
-    "codelyzer": "^6.0.0",
+    "@types/node": "~14.14.10",
+    "codelyzer": "^6.0.1",
     "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "~5.1.1",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~5.2.3",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~3.3.1",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
     "protractor": "~7.0.0",
-    "ts-node": "~8.10.2",
+    "ts-node": "~9.0.0",
     "tslint": "~6.1.2",
-    "typescript": "~3.9.7"
+    "typescript": "~4.0.2"
   },
   "description": "An Ionic project"
 }
