@@ -1,14 +1,14 @@
-import { AppPage } from './app.po';
+import { AppPage } from "./app.po";
 
-describe('new App', () => {
-  let page: AppPage;
+describe("new App", () => {
+	let page: AppPage;
 
-  beforeEach(() => {
-    page = new AppPage();
-  });
+	beforeEach(() => {
+		page = new AppPage();
+	});
 
-  it('should be blank', () => {
-    page.navigateTo();
-    expect(page.getParagraphText()).toContain('The world is your oyster.');
-  });
+	it("should be blank", () => {
+		page.navigateTo();
+		expect(page.getParagraphText()).toContain("The world is your oyster.");
+	});
 });
