@@ -1,47 +1,45 @@
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { TestBed, async } from '@angular/core/testing';
+import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
+import { TestBed, waitForAsync } from "@angular/core/testing";
 
-import { Platform } from '@ionic/angular';
-import { SplashScreen } from '@ionic-native/splash-screen/ngx';
-import { StatusBar } from '@ionic-native/status-bar/ngx';
+import { Platform } from "@ionic/angular";
+import { SplashScreen } from "@ionic-native/splash-screen/ngx";
+import { StatusBar } from "@ionic-native/status-bar/ngx";
 
-import { AppComponent } from './app.component';
+import { AppComponent } from "./app.component";
 
-describe('AppComponent', () => {
+describe("AppComponent", () => {
+	let statusBarSpy, splashScreenSpy, platformReadySpy, platformSpy;
 
-  let statusBarSpy, splashScreenSpy, platformReadySpy, platformSpy;
+	beforeEach(waitForAsync(() => {
+		statusBarSpy = jasmine.createSpyObj("StatusBar", ["styleDefault"]);
+		splashScreenSpy = jasmine.createSpyObj("SplashScreen", ["hide"]);
+		platformReadySpy = Promise.resolve();
+		platformSpy = jasmine.createSpyObj("Platform", { ready: platformReadySpy });
 
-  beforeEach(async(() => {
-    statusBarSpy = jasmine.createSpyObj('StatusBar', ['styleDefault']);
-    splashScreenSpy = jasmine.createSpyObj('SplashScreen', ['hide']);
-    platformReadySpy = Promise.resolve();
-    platformSpy = jasmine.createSpyObj('Platform', { ready: platformReadySpy });
+		TestBed.configureTestingModule({
+			declarations: [AppComponent],
+			schemas: [CUSTOM_ELEMENTS_SCHEMA],
+			providers: [
+				{ provide: StatusBar, useValue: statusBarSpy },
+				{ provide: SplashScreen, useValue: splashScreenSpy },
+				{ provide: Platform, useValue: platformSpy },
+			],
+		}).compileComponents();
+	}));
 
-    TestBed.configureTestingModule({
-      declarations: [AppComponent],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA],
-      providers: [
-        { provide: StatusBar, useValue: statusBarSpy },
-        { provide: SplashScreen, useValue: splashScreenSpy },
-        { provide: Platform, useValue: platformSpy },
-      ],
-    }).compileComponents();
-  }));
+	it("should create the app", () => {
+		const fixture = TestBed.createComponent(AppComponent);
+		const app = fixture.debugElement.componentInstance;
+		expect(app).toBeTruthy();
+	});
 
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app).toBeTruthy();
-  });
+	it("should initialize the app", async () => {
+		TestBed.createComponent(AppComponent);
+		expect(platformSpy.ready).toHaveBeenCalled();
+		await platformReadySpy;
+		expect(statusBarSpy.styleDefault).toHaveBeenCalled();
+		expect(splashScreenSpy.hide).toHaveBeenCalled();
+	});
 
-  it('should initialize the app', async () => {
-    TestBed.createComponent(AppComponent);
-    expect(platformSpy.ready).toHaveBeenCalled();
-    await platformReadySpy;
-    expect(statusBarSpy.styleDefault).toHaveBeenCalled();
-    expect(splashScreenSpy.hide).toHaveBeenCalled();
-  });
-
-  // TODO: add more tests!
-
+	// TODO: add more tests!
 });
